AA35 will not do anything anyway, therefore: changed the type of the PSI-BIN bin_width member from float to double.

+++ src/external/MuSR_software/Class_MuSR_PSI/MuSR_td_PSI_bin.h
@@ -82,7 +82,7 @@ class MuSR_td_PSI_bin {
     char     time_start[9];
     char     time_stop[9];
 
-    float    bin_width;
+    double   bin_width;
 
     int      number_histo;
     int      length_histo;
